databse not fetching all the time

+++ backend/app/database.py
@@ -1,19 +1,20 @@
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker, Session
 
+from app.config import FIRST_RUN, LOCALLY
 from app.models.base import Base
 
-#DATABASE_URL = "postgresql+psycopg2://postgres:postgres@postgres:5432/stock_sim"
-DATABASE_URL = "postgresql+psycopg2://postgres:postgres@localhost:5432/stock_sim"
+#DATABASE_URL = "postgresql+psycopg2://postgres:postgres@" +  + ":5432/stock_sim"
+db = "localhost" if LOCALLY else "postgres"
+DATABASE_URL = f"postgresql+psycopg2://postgres:postgres@{db}:5432/stock_sim"
 
 engine = create_engine(DATABASE_URL, echo=False)
 SessionLocal = sessionmaker(bind=engine)
 
-# Usuń wszystkie tabele
-Base.metadata.drop_all(bind=engine)
+if FIRST_RUN:
+    Base.metadata.drop_all(bind=engine)
+    Base.metadata.create_all(bind=engine)
 
-# Tworzymy wszystkie tabele
-Base.metadata.create_all(bind=engine)
 
 # Dependency do FastAPI
 def get_db() -> Session:
